Refactor modeling into modular subpackages and update imports
